include: acpi: actbl2: Fix RQSC rcid/mcid field width to 16-bit
RCID and MCID in the ACPI RQSC table are 16-bit fields per the
spec. Update struct acpi_table_rqsc_fields to use u16 for rcid/mcid
so the in-kernel parser matches the table layout.

+++ include/acpi/actbl2.h
@@ -3375,8 +3375,8 @@ struct acpi_table_rqsc_fields {
 	u8 resv;	//  1
 	u16 length;	//  2
 	u32 reg[3];	// 12
-	u32 rcid;	//  4
-	u32 mcid;	//  4
+	u16 rcid;	//  2
+	u16 mcid;	//  2
 	u16 flags;	//  2
 	u16 nres;	//  2
 	struct acpi_table_rqsc_fields_res res; // 20
